Refactor `XInput` implementation: Introduce `IResponseProto` and `IVoidProto` interfaces, update `IXinputRequest` inheritance, and replace `sealed` with `readonly` structs in response generation.
