--- Src/Extensions/Xcsb.Extension.XInput/IXinputRequest.cs
@@ -1,69 +1,0 @@
-using System;
-using Xcsb.Extension.XInput.Response.Replies;
-
-namespace Xcsb.Extension.XInput;
-
-public interface IXinputRequest
-{
-    GetExtensionVersionReply GetExtensionVersion(ReadOnlySpan<byte> name);
-    void ListInputDevices();
-    void OpenDevice();
-    void CloseDevice(byte deviceId);
-    void SetDeviceMode();
-    void SelectExtensionEvent();
-    void GetSelectedExtensionEvents();
-    void ChangeDeviceDontPropagateList();
-    void GetDeviceDontPropagateList();
-    void GetDeviceMotionEvents();
-    void ChangeKeyboardDevice();
-    void ChangePointerDevice();
-    void GrabDevice();
-    void UngrabDevice(uint time, byte deviceId);
-    void GrabDeviceKey();
-    void UngrabDeviceKey();
-    void GrabDeviceButton();
-    void UngrabDeviceButton();
-    void AllowDeviceEvents();
-    void GetDeviceFocus();
-    void SetDeviceFocus();
-    void GetFeedbackControl();
-    void ChangeFeedbackControl();
-    void GetDeviceKeyMapping();
-    void ChangeDeviceKeyMapping();
-    void GetDeviceModifierMapping();
-    void SetDeviceModifierMapping();
-    void GetDeviceButtonMapping();
-    void SetDeviceButtonMapping();
-    void QueryDeviceState();
-    void SendExtensionEvent();
-    void DeviceBell();
-    void SetDeviceValuators();
-    void GetDeviceControl();
-    void ChangeDeviceControl();
-    void ListDeviceProperties();
-    void ChangeDeviceProperty();
-    void DeleteDeviceProperty();
-    void GetDeviceProperty();
-    void XIQueryPointer();
-    void XIWarpPointer();
-    void XIChangeCursor();
-    void XIChangeHierarchy();
-    void XISetClientPointer();
-    void XIGetClientPointer();
-    void XISelectEvents();
-    void XIQueryVersion();
-    void XIQueryDevice();
-    void XISetFocus();
-    void XIGetFocus();
-    void XIGrabDevice();
-    void XIUngrabDevice();
-    void XIAllowEvents();
-    void XIPassiveGrabDevice();
-    void XIPassiveUngrabDevice();
-    void XIListProperties();
-    void XIChangeProperty();
-    void XIDeleteProperty();
-    void XIGetProperty();
-    void XIGetSelectedEvents();
-    void XIBarrierReleasePointer();
-}
