+++ Src/Extensions/Xcsb.Extension.XInput/Infrastructure/IXinputRequest.cs
@@ -1,0 +1,11 @@
+using System;
+using Xcsb.Extension.XInput.Infrastructure.ResponceProto;
+using Xcsb.Extension.XInput.Infrastructure.VoidProto;
+using Xcsb.Extension.XInput.Response.Replies;
+
+namespace Xcsb.Extension.XInput.Infrastructure;
+
+public interface IXinputRequest : IVoidProto, IResponceProto
+{
+    GetExtensionVersionReply GetExtensionVersion(ReadOnlySpan<byte> name);
+}
