+++ Src/Extensions/Xcsb.Extension.XInput/Infrastructure/ResponceProto/IResponceProto.cs
@@ -1,0 +1,7 @@
+using Xcsb.Extension.XInput.Response.Replies;
+
+namespace Xcsb.Extension.XInput.Infrastructure.ResponceProto;
+
+public interface IResponceProto
+{
+}
