+++ Src/Extensions/Xcsb.Extension.XInput/Infrastructure/VoidProto/IVoidProto.cs
@@ -1,0 +1,8 @@
+using Xcsb.Connection.Response;
+
+namespace Xcsb.Extension.XInput.Infrastructure.VoidProto;
+
+public interface IVoidProto
+{
+
+}
